Agric_NSW_FarmDamsRiverina {GregsOUPR6}	R Documentation

Water in Farm Dams in the Riverina of New South Wales

Description

Matlab generated water volumes in a 4000 cubic metre dam, with and without tree windbreaks to control evaporation, and a merino sheep flock size of 1500 dry sheep equivalents.

Format

csv file with 19313 rows and 3 columns

Details

Day: time variable in daily increments

Without: water volumes in cubic metres without a windbreak

With: water volumes in cubic metres with a windbreak

Author(s)

Tim Capon tim.capon.csiro.au, Helena Clayton helena.clayton@anu.edu.au, Sally Thompson sally.thompson@uwa.edu.au, Greg Hertzler ghertzlerau@gmail.com, Philip Graham phil@gramadvisory.com.au, David Lindemayer david.lindemayer@anu.edu

[Package GregsOUPR6 version 1.3.5.0 Index]