Agric_NSW_SoilHealthHarden {GregsOUPR6}	R Documentation

Soil Health with Stubble Management

Description

CSIRO Harden Long-Term Tillage Experiment

Format

csv file with 30 rows and 5 columns

Details

Year: time variable in sporadic increments

N Burn: kilograms per hectare of mineral soil nitrogen in the soil profile from 0 to 160 millimetres with stubble burning

N Bash: kilograms per hectare of mineral soil nitrogen in the soil profile from 0 to 160 millimetres with stubble bashing

W Burn: millimetres of water in the soil profile from 0 to 160 millimetres with stubble burning

W Bash: millimetres of water in the soil profile from 0 to 160 millimetres with stubble bashing

Source

https://data.csiro.au/collection/csiro:61293?q=population%20data&_st=keyword&_str=210&_si=7

[Package GregsOUPR6 version 1.3.5.0 Index]