Long Term Crop Rotations, South Australia

Waite Permanent Rotation Trial. v4. CSIRO. Data Collection.

Format

csv file with 70 rows and 30 columns

Source

https://doi.org/10.4225/08/55E5165EC0D29

Details

Year: time variable in annual increments

Plt 1-2 WF W: Wheat Fallow rotation, Wheat yields in kilograms per hectare

Plt 3-4 WPe W: Wheat Peas rotation, Wheat yields in kilograms her hectare

Plt 3-4 WPe Pe: Wheat Peas rotation, Pea yield in kilograms per hectare

Plt 5-7 WPF W: Wheat Pasture Fallow rotation, Wheat yield in kilograms per hectare

Plt 5-7 WPF P: Wheat Pasture Fallow rotation, Pasture dry matter in kilograms per hectare

Plt 8-10 WOF W: Wheat Oats Fallow rotation, Wheat yield in kilograms per hectare

Plt 8-10 WOF O: Wheat Oats Fallow rotation, Oat yield in kilograms per hectare

Plt 11-16 WWPPPP W: Wheatx2 Pasturex4 rotation, Wheat yield in kilograms per hectare

Plt 11-16 WWPPPP P: Wheatx2 Pasturex4 rotation, Pasture dry matter in kilograms per hectare

Plt 17 W W: continuous Wheat, Wheat yield in kilograms per hectare

Plt 18-20 WPP W: Wheat Pasturex2 rotation, Wheat yield in kilograms per hectare

Plt 18-20 WPP P: Wheat Pasturex2 rotation, Pasture dry matter in kilograms per hectare

Plt 21-23 WBPe W: Wheat Barley Peas rotation, Wheat yield in kilograms per hectare

Plt 21-23 WBPe B: Wheat Barley Peas rotation, Barley yield in kilograms per hectare

Plt 21-23 WBPe Pe: Wheat Barley Peas rotation, Pea yield in kilograms per hectare

Plt 24-27 WOPF W: Wheat Oats Pasture Fallow rotation, Wheat yield in kilograms per hectare

Plt 24-27 WOPF O: Wheat Oats Pasture Fallow rotation, Oat yield in kilograms per hectare

Plt 24-27 WOPF P: Wheat Oats Pasture Fallow rotation, Pasture dry matter in kilograms per hectare

Plt 28-29 P P: continuous Pasture, Pasture dry matter in kilograms per hectare

Plt 30-33 WPPF W: Wheat Pasturex2 Fallow rotation, Wheat yield in kilograms per hectare

Plt 30-33 WPPF W: Wheat Pasturex2 Fallow rotation, Pasture dry matter in kilograms per hectare

Plt 34-35 WF W: Wheat Fallow rotation, Wheat yield in kilograms per hectare

Annual Rain: Annual rainfall in millimetres

Apr-Oct Rain: April through October rainfall in millimetres

Plt 1 WF Carbon: Wheat Fallow rotation, Carbon in milligrams per hectare

Plt 10 WOF Carbon: Wheat Oats Fallow rotation, Carbon in milligrams per hectare

Plt 13 WWPPPP Carbon: Wheatx2 Pasturex4 rotation, Carbon in milligrams per hectare

Plt 17 W Carbon: continuous Wheat, Carbon in milligrams per hectare

Plt 29 P Carbon: continuous Pasture, Carbon in milligrams per hectare

Experimental crop yields and pasture dry matter for permanent rotations from 1925 to 1993.

Author

Sanderman, Jonathan; David, Rakesh; Moore, Andrew; Keith, Heather; & Farquharson, Ryan, 2015