Agric_Tas_TreeShelterBelts {GregsOUPR6}	R Documentation

Tree shelter belts in Tasmania

Description

CSIRO Perennial Prosperity Project

Format

csv file with 52 rows and 3 columns

Details

Year: time variable in annual increments

With trees: sheep gross margins per paddock with tree shelter belt

Without trees: sheep gross margins per paddock without tree shelter belt

Author(s)

Tim Capon tim.capon@csiro.au, Daniel Mendham daniel.mendham@csiro.au

[Package GregsOUPR6 version 1.3.5.0 Index]