Climate_SeaLevel_PortKembla {GregsOUPR6}	R Documentation

Sea Level at Port Kembla

Description

Bureau of Meteorology Australian Baseline Sea Level Monitoring Project, with measurements smoothed to eliminate variability.

Format

csv file with 7985 rows and 5 columns

Details

Day in 2023: time variable in hourly increments

Sea Level: metres above Tide Gauge Zero

Water Temperature: degrees Celsius

Air Temperature: degrees Celsius

Barometric Pressure: hPa

This is radically smoothed data. Do not use this data for decisions under uncertainty.

Source

http://www.bom.gov.au/oceanography/projects/abslmp/data/data.shtml

[Package GregsOUPR6 version 1.3.5.0 Index]