Ecosys_Kangaroos {GregsOUPR6}	R Documentation

Kangaroo Population and Harvest in South Australia

Description

Government of South Australia, Department of Environment, population surveys and meat processor records of harvest aggregated across all regions of South Australia.

Format

csv file with 46 rows and 7 columns

Details

Year: time variable

Red Pop: Estimated population of Red Kangaroos in 1000 head

Red Harvest: Number of Red Kangaroos commercially harvested in 1000 head

Grey Pop: Estimated population of Western Grey Kangaroos in 1000 head

Grey Harvest: Number of Western Grey Kangaroos commercially harvested in 1000 head

Euro Pop: Estimated population of European Kangaroos in 1000 head

Euro Harvest: Number of European Kangaroos commercially harvested in 1000 head

Source

https://www.environment.sa.gov.au/topics/animals-and-plants/sustainable-use-of-animals-and-plants/kangaroo-conservation-and-management/quotas-harvest-data

[Package GregsOUPR6 version 1.3.5.0 Index]