Ecosys_TropicalRockLobsters {GregsOUPR6}	R Documentation

Tropical Rock Lobster in Australian Waters

Description

CSIRO Fisheries Biomass Provisioning Services

Format

csv file with 22 rows and 4 columns

Details

Fin Year: time variable for financial year 1 July to 30 June

Catch: kilograms caught per year

GVP: Gross value product in dollars as market price times catch

EV: Exchange value in dollars as the value of ecosystem services as if a market existed

Source

https://doi.org/10.25919/nzrp-6702

[Package GregsOUPR6 version 1.3.5.0 Index]