Finance_Commodities {GregsOUPR6}	R Documentation

Metals and Energy Commodities

Description

Closing prices on the first day of each month from 1 January 2009 to 1 May 2025.

Format

csv file with 197 rows and 8 columns

Details

Day: time variable in days since 1 January 2009

Gold: London Bullion Market fix at the Perth (Aus) Mint in USD/troy ounce

Silver: London Bullion Market fix in USD/troy ounce

Copper: London Metals Exchange spot price, Copper, Grade A, USD/tonne

Iron Ore: Cleared for Export Tianjin port, Iron Ore Fines 62% FE spot price in USD/tonne

WTI: West Texas Intermediate in USD/bbl

Brent: Brent Crude in USD/bbl

Year: time variable as decimal year

Source

https://files.marketindex.com.au/files/workbooks/commodities-workbook.xlsx

[Package GregsOUPR6 version 1.3.5.0 Index]