OUP_SampleSize {GregsOUPR6}	R Documentation

Sample sizes for the Ornstein-Uhlenbeck Process

Description

Monte-Carlo simulation to demonstrate the effect of sample size on discovering the parameters of the Ornstein-Uhlenbeck Process.

Format

csv file with 501 rows and 4 columns

Details

year: time variable in annual increments for all sample paths

small: 5 observations

medium: 50 observations

large: 500 observations

The first 5 and the first 50 observations are the same as those of 500 simulated observations. More observations get closer to the true parameters of rho=0.5, mu=15 and sigma=50.

[Package GregsOUPR6 version 1.3.5.0 Index]