Analytical {GregsOUPR6}	R Documentation

R6 Class implementing Analytical formulas for an Ornstein-Uhlenbeck Process

Description

A set of functions to calculate probabilities, option prices, visiting times, first passage times and decision thresholds–everything for simple exit and entry options in a Real Options Analysis.

Formulas:

z stochastic
  Drift
  Diffusion
y stochastic
  Mean
  MeanToConverge
  Variance
  VarianceToConverge
  Density
  Probability
  DoubleIntegral
x stochastic
  Option
  OptionEnvelope
  DecisionThreshold
  Obligation
t stochastic
  PassageTimeMode
  PassageTimeMedian
  PassateTimeMean
  PassageTimeVariance
  PassageTimePercentiles
  PassageTimeDensity
  PassageTimeProbability

Plots

  PlotDrift
  PlotDiffusion
  PlotMean
  PlotMeanToConverge
  PlotVariance
  PlotVarianceToConverge
  PlotDensity
  PlotProbability
  PlotDoubleIntegral
  PlotOption
  PlotOptionEnvelope
  PlotDecisionThreshold
  PlotObligation
  PlotPassageTimeModeMedianMean
  PlotPassageTimeVariance
  PlotPassageTimePercentiles
  PlotPassageTimeDensity
  PlotPassageTimeProbability

Arguments of functions:

  All arguments are optional in all functions.
OUP parameters
  rho:    rate parameter 0<=rho<inf
  mu:     location parameter -inf<mu<inf
  sigma:  scale parameter -inf<sigma<inf
z stochastic
  z:      vector of states -inf<z<inf
y stochastic
  t:      vector of times s<=t<inf
  y:      vector of states -inf<y<inf
  s:      initial time -inf<s<inf
  x:      initial state -inf<x<inf
  psi:    <=0 for integral -inf to y,
           >0 for integral y to inf
  eps:    proportion remaining after convergence 0<=eps<=1
x stochastic
  s:      vector of times -inf<s<t
  x:      vector of states -inf<x<inf
  t:      terminal time -inf<t<inf
  y:      terminal state -inf<y<inf
  r:      discount rate -inf<r<inf
  phi:    <=0 for exit option,
           >0 for entry option
t stochastic
  t:      vector of times s<=t<inf
  k:      decision threshold -inf<k<int
  s:      initial time -inf<s<inf
  x:      initial state -inf<x<inf
  omega:  degree of irreversibility 0<=omega<=1

Using the formulas:

Demonstration scripts are files in the 'demo' directory. Identify a formula and in the console type:

  demo(A_FormulaName), or
  demo(A_PlotFormulaName).

Discussion:

Mathematical formulas and convergence criteria assume exact arithmetic. Floating-point arithmetic is another matter. It may overflow, underflow or have extreme cancellation. In principle, everything can be calculated by Monte Carlo simulation or Finite Difference methods. These solve continuous equations at discrete nodes, creating another source of error. Monte Carlo simulation of passage times can be accurate or biased, sometimes wildly so. The Finite Difference method can be accurate or give reasonable looking solutions that diverge from the true answers.

Even if great care is taken, it is a good idea to calculate a solution in more than one way. This is one reason for analytical formulas.

Another reason is to speed the calculations. Compared with the Finite Difference method, an analytical formula can calculate an option price up to 10,000 times quicker. Compared with Monte Carlo simulation, an analytical formula can calculate passage time probabilities up to 100,000 times quicker.

The bottom line of a real options analysis is the decision threshold and the mode, median and mean passage times. The many functions available are used in the calculations, but to go straight to the bottom line:

  A <- Analytical$new()
  A$DecisionThreshold()
  A$PassageTimeMean()
  A$PassageTimeVariance()
  A$PlotPassageTimePercentiles(type=1)
  A$PlotPassageTimePercentiles(type=2)

Perhaps a better measure of passage times are the percentiles. The mean does not exist if rho is zero and the variance does not exist if rho is small. The mode and median always exist. Percentiles, which include the median, always exist.

The passage time calculations are challenging. For example, the passage time density is too complicated to explain in this short discussion. But don't be surprised if it is bi-modal or even negative. Care must be taken in interpreting passage times.

Examples in R6 don't work the same way as other R modules. There is only one example for an R6 object, not one for each function in the object. To run examples, the devtools::run_examples() works, but the R command example("Analytical") doesn't. You can copy commands to the clipboard, paste into the console and press Enter. Examples in this help and a simple example at the bottom can be run in this way.

A better alternative is demo(), but this works sometimes, sometimes not. If you don't see the demos for this package, go to Files and demo. You will see the demo names and then type something like demo(A_DecisionThreshold).

Methods

Public methods

Analytical$new()

Analytical$set_oup_params()

Analytical$set_z_stoch_args()

Analytical$set_y_stoch_args()

Analytical$set_x_stoch_args()

Analytical$set_t_stoch_args()

Analytical$set_plot_info()

Analytical$get_all()

Analytical$get_oup_params()

Analytical$get_z_stoch_args()

Analytical$get_y_stoch_args()

Analytical$get_x_stoch_args()

Analytical$get_t_stoch_args()

Analytical$get_plot_info()

Analytical$get_plot_colors()

Analytical$axes_z_stoch()

Analytical$axes_y_stoch()

Analytical$axes_x_stoch()

Analytical$axes_t_stoch()

Analytical$sync_zyxt_stoch()

Analytical$undo_clear()

Analytical$undo_save()

Analytical$undo_undo()

Analytical$Drift()

Analytical$Diffusion()

Analytical$Mean()

Analytical$MeanToConverge()

Analytical$Variance()

Analytical$VarianceToConverge()

Analytical$Density()

Analytical$Probability()

Analytical$DoubleIntegral()

Analytical$Option()

Analytical$OptionEnvelope()

Analytical$DecisionThreshold()

Analytical$Obligation()

Analytical$PassageTimeMode()

Analytical$PassageTimeMedian()

Analytical$PassageTimeMean()

Analytical$PassageTimeVariance()

Analytical$PassageTimePercentiles()

Analytical$PassageTimeDensity()

Analytical$PassageTimeProbability()

Analytical$PlotDrift()

Analytical$PlotDiffusion()

Analytical$PlotMean()

Analytical$PlotMeanToConverge()

Analytical$PlotVariance()

Analytical$PlotVarianceToConverge()

Analytical$PlotDensity()

Analytical$PlotProbability()

Analytical$PlotDoubleIntegral()

Analytical$PlotOption()

Analytical$PlotOptionEnvelope()

Analytical$PlotDecisionThreshold()

Analytical$PlotObligation()

Analytical$PlotPassageTimeModeMedianMean()

Analytical$PlotPassageTimeVariance()

Analytical$PlotPassageTimePercentiles()

Analytical$PlotPassageTimeDensity()

Analytical$PlotPassageTimeProbability()

Method new()

Create an Analytical object

Usage

Analytical$new(OUP = NULL)

Arguments

OUP

pointer set by the OUProcess object

Returns

A new Analytical object

Examples

  A <- Analytical$new()
  A$DecisionThreshold()

Method set_oup_params()

Set OUP parameters

Usage

Analytical$set_oup_params(rho = NULL, mu = NULL, sigma = NULL, who = NULL)

Arguments

rho

rate parameter 0<=rho<inf

mu

location parameter -inf<mu<inf

sigma

scale parameter -inf<sigma<inf

who

identifier for object sending the parameters

Returns

list(rho,mu,sigma)

Method set_z_stoch_args()

Set z as a stochastic state

Usage

Analytical$set_z_stoch_args(z = NULL)

Arguments

z

vector of n states -inf<z<inf

Returns

list(z)

Method set_y_stoch_args()

Set y as a stochastic state and its arguments

Usage

Analytical$set_y_stoch_args(
  t = NULL,
  y = NULL,
  s = NULL,
  x = NULL,
  psi = NULL,
  eps = NULL
)

Arguments

t

vector of m times s<=t<inf

y

vector of n states -inf<y<inf

s

initial time -inf<s<inf

x

initial state -inf<x<inf

psi

<=0 for integral -inf to y, >0 for integral y to inf

eps

proportion remaining after convergence 0<=eps<=1

Returns

list(t,y,s,x,psi,eps)

Method set_x_stoch_args()

Set x as a stochastic state and its arguments

Usage

Analytical$set_x_stoch_args(
  s = NULL,
  x = NULL,
  t = NULL,
  y = NULL,
  r = NULL,
  phi = NULL,
  b = NULL,
  c = NULL,
  who = NULL
)

Arguments

s

vector of m times -inf<s<t

x

vector of n states -inf<x<inf

t

terminal time -inf<t<inf

y

terminal state -inf<y<inf

r

discount rate -inf<r<inf

phi

<=0 for exit option, >0 for entry option

b

lump-sum benefit for entry option

c

lump-sum cost for exit option

who

identifier for object sending the parameters

Returns

list(s,x,t,y,r,phi)

Method set_t_stoch_args()

Set t stochastic arguments

Usage

Analytical$set_t_stoch_args(
  t = NULL,
  k = NULL,
  s = NULL,
  x = NULL,
  z = NULL,
  omega = NULL,
  Ppct = NULL
)

Arguments

t

vector of m times s<=t<inf

k

decision threshold -inf<k<inf

s

initial time -inf<s<inf

x

initial state -inf<x<inf

z

vector of n alternate initial states -inf<z<inf

omega

degree of irreversibility 0<=omega<=1

Ppct

passage time probability for a percentile 0.01<=Ppct<=0.99

Returns

list(t,k,s,x,z,omega)

Method set_plot_info()

Set information for plotting

Usage

Analytical$set_plot_info(
  type = NULL,
  pmax = NULL,
  ptmax = NULL,
  fontfamily = NULL,
  fontsize = NULL,
  fileformat = NULL,
  filewidth = NULL,
  fileheight = NULL,
  theme = NULL,
  opaque = NULL,
  walls = NULL,
  floor = NULL,
  labels = NULL,
  who = NULL
)

Arguments

type

= 1 and 2 for 2D, 3 and 4 for 3D

pmax

maximum vertical axis for transition density plot

ptmax

maximum vertical axis for passage time density plot

fontfamily

font family for plot labels

fontsize

font size for plot labels

fileformat

'png' or 'svg'

filewidth

pixel width of 2D plot, pixel width and height of 3D plot

fileheight

pixel height of 2D plot

theme

'light' or 'dark'

opaque

transparent to opaque background 0.0<=opaque<=1.0

walls

3D walls TRUE or FALSE

floor

3D floor TRUE or FALSE

labels

title and parameters TRUE or FALSE

who

identifier for object sending the parameters

Returns

list(type,font,file,theme,3D)

Method get_all()

Get all arguments, time series and information

Usage

Analytical$get_all()

Returns

list(oup_params,z_stoch_args,y_stoch_args,x_stoch_args,t_stoch_args,plot_info)

Method get_oup_params()

Get OUP parameters

Usage

Analytical$get_oup_params()

Returns

list(rho,mu,sigma)

Method get_z_stoch_args()

Get z as a stochastic state

Usage

Analytical$get_z_stoch_args()

Returns

list(z)

Method get_y_stoch_args()

Get y as a stochastic state and its arguments

Usage

Analytical$get_y_stoch_args()

Returns

list(t,y,s,x,psi,eps)

Method get_x_stoch_args()

Get x as a stochastic state and its arguments

Usage

Analytical$get_x_stoch_args()

Returns

list(s,x,t,y,r,phi)

Method get_t_stoch_args()

Get t stochastic arguments

Usage

Analytical$get_t_stoch_args()

Returns

list(t,k,s,x,z,omega,Ppct)

Method get_plot_info()

Get information for plotting

Usage

Analytical$get_plot_info()

Returns

list(type,font,file,theme,3D,labels)

Method get_plot_colors()

Get colors for plotting

Usage

Analytical$get_plot_colors()

Returns

list(red,ylw,grn,cyn,blu,mgn,gry,background,font,reverse)

Method axes_z_stoch()

Scale axes for z stochastic arguments

Usage

Analytical$axes_z_stoch()

Returns

NULL

Method axes_y_stoch()

Scale axes for y stochastic arguments

Usage

Analytical$axes_y_stoch()

Returns

NULL

Method axes_x_stoch()

Scale axes for x stochastic arguments

Usage

Analytical$axes_x_stoch()

Returns

NULL

Method axes_t_stoch()

Scale axes for t stochastic arguments

Usage

Analytical$axes_t_stoch()

Returns

NULL

Method sync_zyxt_stoch()

Synchronize states

Usage

Analytical$sync_zyxt_stoch()

Returns

NULL

Method undo_clear()

Clear undo list and save current arguments to list

Usage

Analytical$undo_clear()

Returns

1

Method undo_save()

Save current arguments to undo list

Usage

Analytical$undo_save()

Returns

number of argument sets

Method undo_undo()

Replace current arguments from undo list

Usage

Analytical$undo_undo()

Returns

c(index of this argument set, number of argument sets)

Method Drift()

Calculate, plot and return drifts

Usage

Analytical$Drift(z = NULL, rho = NULL, mu = NULL, plotit = TRUE)

Arguments

z

vector of n states

rho

rate parameter 0<=rho<inf

mu

location parameter -inf<mu<inf

plotit

TRUE or FALSE

Returns

list(g(1xn))

Method Diffusion()

Calculate, plot and return diffusions

Usage

Analytical$Diffusion(z = NULL, sigma = NULL, plotit = TRUE)

Arguments

z

vector of n states

sigma

scale parameter -inf<sigma<inf

plotit

TRUE or FALSE

Returns

list(h2(1xn))

Method Mean()

Calculate and plot means

Usage

Analytical$Mean(
  t = NULL,
  s = NULL,
  x = NULL,
  rho = NULL,
  mu = NULL,
  plotit = TRUE
)

Arguments

t

vector of m times s<=t<inf

s

initial time -inf<s<inf

x

initial state -inf<x<inf

rho

rate parameter 0<=rho<inf

mu

location parameter -inf<mu<inf

plotit

TRUE or FALSE

Returns

list(G(mx1))

Method MeanToConverge()

Calculate and plot time for mean to converge

Usage

Analytical$MeanToConverge(s = NULL, rho = NULL, eps = NULL, plotit = TRUE)

Arguments

s

initial time -inf<s<inf

rho

rate parameter 0<=rho<inf

eps

proportion remaining after convergence 0<=eps<=1

plotit

TRUE or FALSE

Returns

list(Gteps)

Method Variance()

Calculate and plot variances

Usage

Analytical$Variance(
  t = NULL,
  s = NULL,
  rho = NULL,
  sigma = NULL,
  plotit = TRUE
)

Arguments

t

vector of m times s<=t<inf

s

initial time -inf<s<inf

rho

rate parameter 0<=rho<inf

sigma

scale parameter -inf<sigma<inf

plotit

TRUE or FALSE

Returns

list(H2(mx1))

Method VarianceToConverge()

Calculate and plot time for variance to converge

Usage

Analytical$VarianceToConverge(s = NULL, rho = NULL, eps = NULL, plotit = TRUE)

Arguments

s

initial time -inf<s<inf

rho

rate parameter 0<=rho<inf

eps

proportion remaining after convergence 0<=eps<=1

plotit

TRUE or FALSE

Returns

list(H2teps)

Method Density()

Calculate and plot transition densities

Usage

Analytical$Density(
  t = NULL,
  y = NULL,
  s = NULL,
  x = NULL,
  rho = NULL,
  mu = NULL,
  sigma = NULL,
  plotit = TRUE
)

Arguments

t

vector of m times s<=t<inf

y

vector of n states -inf<y<inf

s

initial time -inf<s<inf

x

initial state -inf<x<inf

rho

rate parameter 0<=rho<inf

mu

location parameter -inf<mu<inf

sigma

scale parameter -inf<sigma<inf

plotit

TRUE or FALSE

Returns

list(p(mxn))

Method Probability()

Calculate and plot transition probabilities

Usage

Analytical$Probability(
  t = NULL,
  y = NULL,
  s = NULL,
  x = NULL,
  rho = NULL,
  mu = NULL,
  sigma = NULL,
  psi = NULL,
  plotit = TRUE
)

Arguments

t

vector of m times s<=t<inf

y

vector of nstates -inf<y<inf

s

initial time -inf<s<inf

x

initial state -inf<x<inf

rho

rate parameter 0<=rho<inf

mu

location parameter -inf<mu<inf

sigma

scale parameter -inf<sigma<inf

psi

<=0 for integral -inf to y, >0 for integral y to inf

plotit

TRUE or FALSE

Returns

list(P(mxn))

Method DoubleIntegral()

Calculate and plot double integrals of transition densities

Usage

Analytical$DoubleIntegral(
  t = NULL,
  y = NULL,
  s = NULL,
  x = NULL,
  rho = NULL,
  mu = NULL,
  sigma = NULL,
  psi = NULL,
  plotit = TRUE
)

Arguments

t

vector of m times s<=t<inf

y

vector of n states -inf<y<inf

s

initial time -inf<s<inf

x

initial state -inf<x<inf

rho

rate parameter 0<=rho<inf

mu

location parameter -inf<mu<inf

sigma

scale parameter -inf<sigma<inf

psi

<=0 for integral -inf to y, >0 for integral y to inf

plotit

TRUE or FALSE

Returns

list(PP(mxn))

Method Option()

Calculate and plot option prices

Usage

Analytical$Option(
  s = NULL,
  x = NULL,
  t = NULL,
  y = NULL,
  rho = NULL,
  mu = NULL,
  sigma = NULL,
  r = NULL,
  phi = NULL,
  b = NULL,
  c = NULL,
  plotit = TRUE
)

Arguments

s

vector of m times -inf<s<t

x

vector of n states -inf<x<inf

t

terminal time -inf<t<inf

y

terminal state -inf<y<inf

rho

rate parameter 0<=rho<inf

mu

location parameter -inf<mu<inf

sigma

scale parameter -inf<sigma<inf

r

discount rate -inf<r<inf

phi

<=0 for exit option, >0 for entry option

b

lump-sum benefit for entry option

c

lump-sum cost for exit option

plotit

TRUE or FALSE

Returns

list(O(mxn))

Method OptionEnvelope()

Calculate and plot the envelope of option prices

Usage

Analytical$OptionEnvelope(
  x = NULL,
  y = NULL,
  rho = NULL,
  mu = NULL,
  sigma = NULL,
  r = NULL,
  phi = NULL,
  b = NULL,
  c = NULL,
  plotit = TRUE
)

Arguments

x

vector of n states -inf<x<inf

y

terminal state -inf<y<inf

rho

rate parameter 0<=rho<inf

mu

location parameter -inf<mu<inf

sigma

scale parameter -inf<sigma<inf

r

discount rate -inf<r<inf

phi

<=0 for exit option, >0 for entry option

b

lump-sum benefit for entry option

c

lump-sum cost for exit option

plotit

TRUE or FALSE

Returns

list(Ohat(1xn),shat(1xn))

Method DecisionThreshold()

Calculate and plot the decision threshold

Usage

Analytical$DecisionThreshold(
  x = NULL,
  y = NULL,
  rho = NULL,
  mu = NULL,
  sigma = NULL,
  r = NULL,
  phi = NULL,
  b = NULL,
  c = NULL,
  plotit = TRUE
)

Arguments

x

vector of n states -inf<x<inf

y

terminal state -inf<y<inf

rho

rate parameter 0<=rho<inf

mu

location parameter -inf<mu<inf

sigma

scale parameter -inf<sigma<inf

r

discount rate -inf<r<inf

phi

<=0 for exit option, >0 for entry option

b

lump-sum benefit for entry option

c

lump-sum cost for exit option

plotit

TRUE or FALSE

Returns

list(k,O)

Method Obligation()

Calculate and plot obligations and prohibitions, ie. benefit/cost analysis

Usage

Analytical$Obligation(
  s = NULL,
  x = NULL,
  t = NULL,
  y = NULL,
  rho = NULL,
  mu = NULL,
  r = NULL,
  phi = NULL,
  b = NULL,
  c = NULL,
  plotit = TRUE
)

Arguments

s

vector of m times -inf<s<t

x

vector of n states -inf<x<inf

t

terminal time -inf<t<inf

y

terminal state -inf<y<inf

rho

rate parameter 0<=rho<inf

mu

location parameter -inf<mu<inf

r

discount rate -inf<r<inf

phi

<=0 for obligation, >0 for prohibition

b

lump-sum benefit for entry option

c

lump-sum cost for exit option

plotit

TRUE or FALSE

Returns

list(BC(mxn))

Method PassageTimeMode()

Calculate and plot passage time modes

Usage

Analytical$PassageTimeMode(
  k = NULL,
  s = NULL,
  x = NULL,
  z = NULL,
  omega = NULL,
  rho = NULL,
  mu = NULL,
  sigma = NULL,
  plotit = TRUE
)

Arguments

k

decision threshold -inf<k<int

s

initial time -inf<s<inf

x

initial state -inf<x<inf

z

vector of alternate initial states -inf<x<inf

omega

degree of irreversibility 0<=omega<=1

rho

rate parameter 0<=rho<inf

mu

location parameter -inf<mu<inf

sigma

scale parameter -inf<sigma<inf

plotit

TRUE or FALSE

Returns

list(tmode(1xn),tmodes(mxn))

Method PassageTimeMedian()

Calculate and plot passage time medians

Usage

Analytical$PassageTimeMedian(
  k = NULL,
  s = NULL,
  x = NULL,
  z = NULL,
  omega = NULL,
  rho = NULL,
  mu = NULL,
  sigma = NULL,
  plotit = TRUE
)

Arguments

k

decision threshold -inf<k<int

s

initial time -inf<s<inf

x

initial state -inf<x<inf

z

vector of alternate initial states -inf<x<inf

omega

degree of irreversibility 0<=omega<=1

rho

rate parameter 0<=rho<inf

mu

location parameter -inf<mu<inf

sigma

scale parameter -inf<sigma<inf

plotit

TRUE or FALSE

Returns

list(tmedian(1xn),tmedians(mxn))

Method PassageTimeMean()

Calculate and plot passage time means

Usage

Analytical$PassageTimeMean(
  k = NULL,
  s = NULL,
  x = NULL,
  z = NULL,
  omega = NULL,
  rho = NULL,
  mu = NULL,
  sigma = NULL,
  plotit = TRUE
)

Arguments

k

decision threshold -inf<k<int

s

initial time -inf<s<inf

x

initial state -inf<x<inf

z

vector of alternate initial states -inf<x<inf

omega

degree of irreversibility 0<=omega<=1

rho

rate parameter 0<=rho<inf

mu

location parameter -inf<mu<inf

sigma

scale parameter -inf<sigma<inf

plotit

TRUE or FALSE

Returns

list(tmean(1xn),tmeans(mxn))

Method PassageTimeVariance()

Calculate and plot passage time variances

Usage

Analytical$PassageTimeVariance(
  k = NULL,
  s = NULL,
  x = NULL,
  z = NULL,
  omega = NULL,
  rho = NULL,
  mu = NULL,
  sigma = NULL,
  plotit = TRUE
)

Arguments

k

decision threshold -inf<k<int

s

initial time -inf<s<inf

x

initial state -inf<x<inf

z

vector of alternate initial states -inf<x<inf

omega

degree of irreversibility 0<=omega<=1

rho

rate parameter 0<=rho<inf

mu

location parameter -inf<mu<inf

sigma

scale parameter -inf<sigma<inf

plotit

TRUE or FALSE

Returns

(tvariance(1xn),tvariances(mxn))

Method PassageTimePercentiles()

Calculate and plot passage time percentiles

Usage

Analytical$PassageTimePercentiles(
  k = NULL,
  s = NULL,
  x = NULL,
  z = NULL,
  omega = NULL,
  Ppct = NULL,
  rho = NULL,
  mu = NULL,
  sigma = NULL,
  plotit = TRUE
)

Arguments

k

decision threshold -inf<k<int

s

initial time -inf<s<inf

x

initial state -inf<x<inf

z

vector of alternate initial states -inf<x<inf

omega

degree of irreversibility 0<=omega<=1

Ppct

passage time probability for a percentile 0.01<=Ppct<=0.99

rho

rate parameter 0<=rho<inf

mu

location parameter -inf<mu<inf

sigma

scale parameter -inf<sigma<inf

plotit

TRUE or FALSE

Returns

list(tpercentile(1xn),tpercentiles(mxn))

Method PassageTimeDensity()

Calculate and plot passage time densities

Usage

Analytical$PassageTimeDensity(
  t = NULL,
  k = NULL,
  s = NULL,
  x = NULL,
  z = NULL,
  omega = NULL,
  rho = NULL,
  mu = NULL,
  sigma = NULL,
  plotit = TRUE
)

Arguments

t

vector of m times s<=t<inf

k

decision threshold -inf<k<int

s

initial time -inf<s<inf

x

initial state -inf<x<inf

z

vector of alternate initial states -inf<x<inf

omega

degree of irreversibility 0<=omega<=1

rho

rate parameter 0<=rho<inf

mu

location parameter -inf<mu<inf

sigma

scale parameter -inf<sigma<inf

plotit

TRUE or FALSE

Returns

list(ptx(1xn),pt(mxn))

Method PassageTimeProbability()

Calculate and plot passage time probabilities

Usage

Analytical$PassageTimeProbability(
  t = NULL,
  k = NULL,
  s = NULL,
  x = NULL,
  z = NULL,
  omega = NULL,
  rho = NULL,
  mu = NULL,
  sigma = NULL,
  plotit = TRUE
)

Arguments

t

vector of m times s<=t<inf

k

decision threshold -inf<k<int

s

initial time -inf<s<inf

x

initial state -inf<x<inf

z

vector of alternate initial states -inf<x<inf

omega

degree of irreversibility 0<=omega<=1

rho

rate parameter 0<=rho<inf

mu

location parameter -inf<mu<inf

sigma

scale parameter -inf<sigma<inf

plotit

TRUE or FALSE

Returns

list(Ptx(1xn),Pt(mxn))

Method PlotDrift()

Plot drifts

Usage

Analytical$PlotDrift(
  title = NULL,
  xaxis = NULL,
  yaxis = NULL,
  zbeg = NULL,
  zend = NULL
)

Arguments

title

text for plot title

xaxis

text for x-axis label

yaxis

text for y-axis label

zbeg

begin value for state axis

zend

end value for state axis

Returns

plot

Method PlotDiffusion()

Plot diffusions

Usage

Analytical$PlotDiffusion(
  type = NULL,
  title = NULL,
  xaxis = NULL,
  yaxis = NULL,
  zbeg = NULL,
  zend = NULL
)

Arguments

type

= 2 and 3 for 2D

title

text for plot title

xaxis

text for x-axis label

yaxis

text for y-axis label

zbeg

begin value for state axis

zend

end value for state axis

Returns

plot

Method PlotMean()

Plot means

Usage

Analytical$PlotMean(
  type = NULL,
  pmax = NULL,
  title = NULL,
  xaxis = NULL,
  yaxis = NULL,
  zaxis = NULL,
  tbeg = NULL,
  tend = NULL,
  ybeg = NULL,
  yend = NULL
)

Arguments

type

= 3 for 2D, 4 and 5 for 3D

pmax

maximum scale for vertical axis

title

text for plot title

xaxis

text for x-axis label

yaxis

text for y-axis label

zaxis

text for z-axis label

tbeg

begin value for time axis

tend

end value for time axis

ybeg

begin value for state axis

yend

end value for state axis

Returns

plot

Method PlotMeanToConverge()

Plot the mean to converge

Usage

Analytical$PlotMeanToConverge(
  title = NULL,
  xaxis = NULL,
  yaxis = NULL,
  tbeg = NULL,
  tend = NULL
)

Arguments

title

text for plot title

xaxis

text for x-axis label

yaxis

text for y-axis label

tbeg

begin value for time axis

tend

end value for time axis

Returns

plot

Method PlotVariance()

Plot variances

Usage

Analytical$PlotVariance(
  type = NULL,
  pmax = NULL,
  title = NULL,
  xaxis = NULL,
  yaxis = NULL,
  zaxis = NULL,
  tbeg = NULL,
  tend = NULL,
  ybeg = NULL,
  yend = NULL
)

Arguments

type

= 2 and 3 for 2D, 4 and 5 for 3D

pmax

maximum scale for vertical axis

title

text for plot title

xaxis

text for x-axis label

yaxis

text for y-axis label

zaxis

text for z-axis label

tbeg

begin value for time axis

tend

end value for time axis

ybeg

begin value for state axis

yend

end value for state axis

Returns

plot

Method PlotVarianceToConverge()

Plot the variance to converge

Usage

Analytical$PlotVarianceToConverge(
  title = NULL,
  xaxis = NULL,
  yaxis = NULL,
  tbeg = NULL,
  tend = NULL
)

Arguments

title

text for plot title

xaxis

text for x-axis label

yaxis

text for y-axis label

tbeg

begin value for time axis

tend

end value for time axis

Returns

plot

Method PlotDensity()

Plot transition densities

Usage

Analytical$PlotDensity(
  type = NULL,
  pmax = NULL,
  title = NULL,
  xaxis = NULL,
  yaxis = NULL,
  zaxis = NULL,
  tbeg = NULL,
  tend = NULL,
  ybeg = NULL,
  yend = NULL
)

Arguments

type

= 2 for 2D, 3, 4 and 5 for 3D

pmax

maximum scale for vertical axis

title

text for plot title

xaxis

text for x-axis label

yaxis

text for y-axis label

zaxis

text for z-axis label

tbeg

begin value for time axis

tend

end value for time axis

ybeg

begin value for state axis

yend

end value for state axis

Returns

plot

Method PlotProbability()

Plot transition probabilities

Usage

Analytical$PlotProbability(
  type = NULL,
  title = NULL,
  xaxis = NULL,
  yaxis = NULL,
  zaxis = NULL,
  tbeg = NULL,
  tend = NULL,
  ybeg = NULL,
  yend = NULL
)

Arguments

type

= 2 for 2D, 3, 4 and 5 for 3D

title

text for plot title

xaxis

text for x-axis label

yaxis

text for y-axis label

zaxis

text for z-axis label

tbeg

begin value for time axis

tend

end value for time axis

ybeg

begin value for state axis

yend

end value for state axis

Returns

plot

Method PlotDoubleIntegral()

Plot double integrals of transition densities

Usage

Analytical$PlotDoubleIntegral(
  type = NULL,
  title = NULL,
  xaxis = NULL,
  yaxis = NULL,
  zaxis = NULL,
  tbeg = NULL,
  tend = NULL,
  ybeg = NULL,
  yend = NULL
)

Arguments

type

= 2 for 2D, 3, 4 and 5 for 3D

title

text for plot title

xaxis

text for x-axis label

yaxis

text for y-axis label

zaxis

text for z-axis label

tbeg

begin value for time axis

tend

end value for time axis

ybeg

begin value for state axis

yend

end value for state axis

Returns

plot

Method PlotOption()

Plot options

Usage

Analytical$PlotOption(
  type = NULL,
  title = NULL,
  xaxis = NULL,
  yaxis = NULL,
  zaxis = NULL,
  sbeg = NULL,
  send = NULL,
  xbeg = NULL,
  xend = NULL
)

Arguments

type

= 2 for 2D, 3, 4 and 5 for 3D

title

text for plot title

xaxis

text for x-axis label

yaxis

text for y-axis label

zaxis

text for z-axis label

sbeg

begin value for time axis

send

end value for time axis

xbeg

begin value for state axis

xend

end value for state axis

Returns

plot

Method PlotOptionEnvelope()

Plot the option envelope

Usage

Analytical$PlotOptionEnvelope(
  type = NULL,
  title = NULL,
  xaxis = NULL,
  yaxis = NULL,
  zaxis = NULL,
  sbeg = NULL,
  send = NULL,
  xbeg = NULL,
  xend = NULL
)

Arguments

type

= 3 for 2D, 4, 5 and 6 for 3D

title

text for plot title

xaxis

text for x-axis label

yaxis

text for y-axis label

zaxis

text for z-axis label

sbeg

begin value for time axis

send

end value for time axis

xbeg

begin value for state axis

xend

end value for state axis

Returns

plot

Method PlotDecisionThreshold()

Plot the decision threshold

Usage

Analytical$PlotDecisionThreshold(
  title = NULL,
  xaxis = NULL,
  yaxis = NULL,
  xbeg = NULL,
  xend = NULL
)

Arguments

title

text for plot title

xaxis

text for x-axis label

yaxis

text for y-axis label

xbeg

begin value for state axis

xend

end value for state axis

Returns

plot

Method PlotObligation()

Plot obligations

Usage

Analytical$PlotObligation(
  type = NULL,
  title = NULL,
  xaxis = NULL,
  yaxis = NULL,
  zaxis = NULL,
  sbeg = NULL,
  send = NULL,
  xbeg = NULL,
  xend = NULL
)

Arguments

type

= 2 for 2D, 3, 4, 5 and 6 for 3D

title

text for plot title

xaxis

text for x-axis label

yaxis

text for y-axis label

zaxis

text for z-axis label

sbeg

begin value for time axis

send

end value for time axis

xbeg

begin value for state axis

xend

end value for state axis

Returns

plot

Method PlotPassageTimeModeMedianMean()

Plot passage time modes, medians and means

Usage

Analytical$PlotPassageTimeModeMedianMean(
  type = NULL,
  xaxis = NULL,
  yaxis = NULL,
  zaxis = NULL,
  ptmax = NULL,
  title = NULL,
  tbeg = NULL,
  tend = NULL,
  zbeg = NULL,
  zend = NULL
)

Arguments

type

= 1, 2, 3 and 4 for 2D, 5 and 6 for 3D

xaxis

text for x-axis label

yaxis

text for y-axis label

zaxis

text for z-axis label

ptmax

maximum scale for vertical axis

title

text for plot title

tbeg

begin value for time axis

tend

end value for time axis

zbeg

begin value for state axis

zend

end value for state axis

Returns

plot

Method PlotPassageTimeVariance()

Plot passage time variances

Usage

Analytical$PlotPassageTimeVariance(
  type = NULL,
  xaxis = NULL,
  yaxis = NULL,
  title = NULL,
  zbeg = NULL,
  zend = NULL
)

Arguments

type

= 3 and 4 for 2D

xaxis

text for x-axis label

yaxis

text for y-axis label

title

text for plot title

zbeg

begin value for state axis

zend

end value for state axis

Returns

plot

Method PlotPassageTimePercentiles()

Plot passage time lower, median and upper percentiles

Usage

Analytical$PlotPassageTimePercentiles(
  type = NULL,
  ptmax = NULL,
  title = NULL,
  xaxis = NULL,
  yaxis = NULL,
  zaxis = NULL,
  tbeg = NULL,
  tend = NULL,
  zbeg = NULL,
  zend = NULL
)

Arguments

type

= 1, 2, 3 and 4 for 2D, 5 and 6 for 3D

ptmax

maximum scale for vertical axis

title

text for plot title

xaxis

text for x-axis label

yaxis

text for y-axis label

zaxis

text for z-axis label

tbeg

begin value for time axis

tend

end value for time axis

zbeg

begin value for state axis

zend

end value for state axis

Returns

plot

Method PlotPassageTimeDensity()

Plot passage time densities

Usage

Analytical$PlotPassageTimeDensity(
  type = NULL,
  ptmax = NULL,
  title = NULL,
  xaxis = NULL,
  yaxis = NULL,
  zaxis = NULL,
  tbeg = NULL,
  tend = NULL,
  zbeg = NULL,
  zend = NULL
)

Arguments

type

= 1 and 2 for 2D, 3, 4 and 5 for 3D

ptmax

maximum scale for vertical axis

title

text for plot title

xaxis

text for x-axis label

yaxis

text for y-axis label

zaxis

text for z-axis label

tbeg

begin value for time axis

tend

end value for time axis

zbeg

begin value for state axis

zend

end value for state axis

Returns

plot

Method PlotPassageTimeProbability()

Plot passage time probabilities

Usage

Analytical$PlotPassageTimeProbability(
  type = NULL,
  title = NULL,
  xaxis = NULL,
  yaxis = NULL,
  zaxis = NULL,
  tbeg = NULL,
  tend = NULL,
  zbeg = NULL,
  zend = NULL
)

Arguments

type

= 1 and 2 for 2D, 3, 4 and 5 for 3D

title

text for plot title

xaxis

text for x-axis label

yaxis

text for y-axis label

zaxis

text for z-axis label

tbeg

begin value for time axis

tend

end value for time axis

zbeg

begin value for state axis

zend

end value for state axis

Returns

plot

Examples

## ------------------------------------------------
## Method `Analytical$new`
## ------------------------------------------------

  A <- Analytical$new()
  A$DecisionThreshold()

[Package GregsOUPR6 version 1.3.5.0 Index]