FiniteDifference {GregsOUPR6}	R Documentation

R6 Class implementing a Finite Difference method.

Description

A Finite Difference method for the Ornstein-Uhlenbeck Process–without boundary conditions on the state but with arbitrary terminal values–for the pricing of index insurance, perpetual options and sequential options. From the option prices, an option envelope and a decision threshold can be calculated. Drift, diffusion and several possible terminal values are pre-programmed. User-defined terminal values can also be entered.

Formulas and methods:

x stochastic
  Drift
  Diffusion
  TerminalValue_Linear
  TerminalValue_Degenerate
  TerminalValue_Stepped
  TerminalValue_Kinked
  TerminalValue_Butterfly
  TerminalValue_Mitscherlich
  TerminalValue_Gompertz
  TerminalValue_Logistic
  TerminalValue_Transcendental
  TerminalValue_YieldIndex
  TerminalValue
  Option
  OptionEnvelope
  DecisionThreshold

Plots

  PlotDrift
  PlotDiffusion
  PlotTerminalValue
  PlotOption
  PlotOptionEnvelope
  PlotDecisionThreshold

Arguments of functions:

  All arguments are optional in all functions.
  s:     vector of times
  x:     vector of states
  V:     vector of terminal values
  r:     discount rate
  phi:   search direction for exit or entry options
  theta: weight of current time in time stepping
  skip:  divide the time interval and report every skip result
  rho:   rate parameter
  mu:    location parameter
  sigma: scale parameter
  xo:    state at the intercept, spike, step or kink
  xi:    state at the inflection point
  xm:    state at the maximum or kink
  vs:    slope or direction of a step
  vr:    rate of change
  Vmax:  maximum terminal value
  Vmin:  minimum terminal value

Using the formulas and methods:

Demonstration scripts are in files in the 'demo' directory. Identify a formula and in the console type:

  demo(FD_FormulaName), or
  demo(FD_PlotFormulaName).

Discussion:

The Finite Difference Method can solve problems with no analytical solutions. As a numerical procedure, it will have errors. Large dif fusions will cause larger errors. To manage the errors, the domain for x should be as large as practical. The increment between times is ds and skip subdivides time into small intervals. Only calculations for ds are reported. The increment between states is dx and ds/skip should be at most dx/100. Increasing skip or dx may reduce errors up to a point. If possible, the Finite Difference Method should be calibrated against an analytical solution for a simple problem before solving more complicated problems.

The functions include many different types of terminal values, but suppose the bottom line is a simple exit option:

  FD <- FiniteDifference$new()
  FD$DecisionThreshold(mu=10)

Or a simple entry option:

  FD$TerminalValue_Kinked(xo=5,vs=1)
  FD$DecisionThreshold()

(That's x oh, not x zero). And finally a sequence of an entry with an option to exit, calculating the exit decision first:

  FD$TerminalValue_Kinked(xo=0,vs=-1)
  exit <- FD$OptionEnvelope()$Ohat
  V <- exit+FD$TerminalValue_Kinked(xo=5,vs=1)$V
  FD$DecisionThreshold(V=V,phi=1)

The functions all return named lists. Before being used in calculations, a list must be stripped of its names, the names Ohat for the option envelope and V for the terminal value, in the above example. The functions try to help by accepting named lists as arguments. For example:

  V <- FD$TerminalValue_Kinked(xo=5,vs=1)
  FD$DecisionThreshold(V=V)

But the terminal value functions automatically set the terminal value in the object, so entering V as as argument does nothing useful in this case.

You may wish to calculate your own terminal value in the console. You must first get rid of the names. The names to get rid of are listed under 'Returns' in the help for each function. You can also use double brackets to eliminate names if you know the position in the list. Both 'Ohat' and 'V' are first in their lists:

  FD$TerminalValue_Kinked(xo=0,vs=-1)
  exit <- FD$OptionEnvelope()[[1]]
  V <- exit+FD$TerminalValue_Kinked(xo=5,vs=1)[[1]]
  FD$DecisionThreshold(V=V,phi=1)

Sorry.

Examples in R6 don't work the same way as other R modules. There is only one example for an R6 object, not one for each function in the object. To run examples, the devtools::run_examples() works, but the R command example("FiniteDifference") doesn't. You can copy commands to the clipboard, paste into the console and press Enter. Examples in this help and a simple example at the bottom can be run in this way.

A better alternative is demo(), but this works sometimes, sometimes not. If you don't see the demos for this package, go to Files and demo. You will see the demo names and then type something like demo(FD_DecisionThreshold).

Methods

Public methods

FiniteDifference$new()

FiniteDifference$set_oup_params()

FiniteDifference$set_x_stoch_args()

FiniteDifference$set_V_linear_args()

FiniteDifference$set_V_degenerate_args()

FiniteDifference$set_V_stepped_args()

FiniteDifference$set_V_kinked_args()

FiniteDifference$set_V_butterfly_args()

FiniteDifference$set_V_mitscherlich_args()

FiniteDifference$set_V_gompertz_args()

FiniteDifference$set_V_logistic_args()

FiniteDifference$set_V_transcendental_args()

FiniteDifference$set_V_yieldindex_args()

FiniteDifference$set_V_args()

FiniteDifference$set_V_info()

FiniteDifference$set_plot_info()

FiniteDifference$get_all()

FiniteDifference$get_oup_params()

FiniteDifference$get_x_stoch_args()

FiniteDifference$get_V_linear_args()

FiniteDifference$get_V_degenerate_args()

FiniteDifference$get_V_stepped_args()

FiniteDifference$get_V_kinked_args()

FiniteDifference$get_V_butterfly_args()

FiniteDifference$get_V_mitscherlich_args()

FiniteDifference$get_V_gompertz_args()

FiniteDifference$get_V_logistic_args()

FiniteDifference$get_V_transcendental_args()

FiniteDifference$get_V_yieldindex_args()

FiniteDifference$get_V_args()

FiniteDifference$get_V_info()

FiniteDifference$get_plot_info()

FiniteDifference$get_plot_colors()

FiniteDifference$axes_x_stoch()

FiniteDifference$undo_clear()

FiniteDifference$undo_save()

FiniteDifference$undo_undo()

FiniteDifference$Drift()

FiniteDifference$Diffusion()

FiniteDifference$TerminalValue_Linear()

FiniteDifference$TerminalValue_Degenerate()

FiniteDifference$TerminalValue_Stepped()

FiniteDifference$TerminalValue_Kinked()

FiniteDifference$TerminalValue_Butterfly()

FiniteDifference$TerminalValue_Mitscherlich()

FiniteDifference$TerminalValue_Gompertz()

FiniteDifference$TerminalValue_Logistic()

FiniteDifference$TerminalValue_Transcendental()

FiniteDifference$TerminalValue_YieldIndex()

FiniteDifference$TerminalValue()

FiniteDifference$Option()

FiniteDifference$OptionEnvelope()

FiniteDifference$DecisionThreshold()

FiniteDifference$PlotDrift()

FiniteDifference$PlotDiffusion()

FiniteDifference$PlotTerminalValue()

FiniteDifference$PlotOption()

FiniteDifference$PlotOptionEnvelope()

FiniteDifference$PlotDecisionThreshold()

Method new()

Create a new FiniteDifference object

Usage

FiniteDifference$new(OUP = NULL)

Arguments

OUP

pointer set by the OUProcess object

Returns

A new FiniteDifference object

Examples

  FD <- FiniteDifference$new()
  FD$TerminalValue_Mitscherlich(plotit=FALSE)
  FD$DecisionThreshold()

Method set_oup_params()

Set OUP parameters

Usage

FiniteDifference$set_oup_params(
  rho = NULL,
  mu = NULL,
  sigma = NULL,
  who = NULL
)

Arguments

rho

rate parameter 0<=rho<inf

mu

location parameter -inf<mu<inf

sigma

scale parameter -inf<sigma<inf

who

identifier for object sending the parameters

Returns

list(rho,mu,sigma)

Method set_x_stoch_args()

Set x as a stochastic state and its arguments

Usage

FiniteDifference$set_x_stoch_args(
  s = NULL,
  x = NULL,
  V = NULL,
  r = NULL,
  phi = NULL,
  theta = NULL,
  skip = NULL,
  who = NULL
)

Arguments

s

vector of times

x

vector of states

V

vector of terminal values

r

discount rate 0<=r<inf, scalar

phi

search direction for exit or entry options

theta

weight of current time in time stepping 0.5<=theta<=1

skip

divide the time interval and report every skip result

who

identifier for object sending the parameters

Returns

list(s,x,V,r,phi,theta,skip,ds,dx)

Method set_V_linear_args()

Set V linear arguments

Usage

FiniteDifference$set_V_linear_args(xo = NULL, vs = NULL)

Arguments

xo

state at the intercept

vs

slope

Returns

list(xo,vs)

Method set_V_degenerate_args()

Set V degenerate arguments

Usage

FiniteDifference$set_V_degenerate_args(xo = NULL, Vmax = NULL, Vmin = NULL)

Arguments

xo

state at the step

Vmax

maximum terminal value

Vmin

minimum terminal value

Returns

list(xo,vs,Vmax,Vmin)

Method set_V_stepped_args()

Set V stepped arguments

Usage

FiniteDifference$set_V_stepped_args(
  xo = NULL,
  vs = NULL,
  Vmax = NULL,
  Vmin = NULL
)

Arguments

xo

state at the step

vs

direction of the step

Vmax

maximum terminal value

Vmin

minimum terminal value

Returns

list(xo,vs,Vmax,Vmin)

Method set_V_kinked_args()

Set V kinked arguments

Usage

FiniteDifference$set_V_kinked_args(
  xo = NULL,
  vs = NULL,
  Vmax = NULL,
  Vmin = NULL
)

Arguments

xo

state at the kink

vs

slope

Vmax

maximum terminal value

Vmin

minimum terminal value

Returns

list(xo,vs,Vmax,Vmin)

Method set_V_butterfly_args()

Set V butterfly arguments

Usage

FiniteDifference$set_V_butterfly_args(
  xo = NULL,
  xm = NULL,
  vs = NULL,
  Vmax = NULL,
  Vmin = NULL
)

Arguments

xo

state at the left wing

xm

state at the right wing

vs

slope

Vmax

maximum terminal value

Vmin

minimum terminal value

Returns

list(xo,xm,vs,Vmax,Vmin)

Method set_V_mitscherlich_args()

Set V mitscherlich arguments

Usage

FiniteDifference$set_V_mitscherlich_args(
  xo = NULL,
  vr = NULL,
  Vmax = NULL,
  Vmin = NULL
)

Arguments

xo

state at the intercept

vr

rate of change

Vmax

maximum terminal value

Vmin

minimum terminal value

Returns

list(xo,vr,Vmax,Vmin)

Method set_V_gompertz_args()

Set V gompertz arguments

Usage

FiniteDifference$set_V_gompertz_args(
  xi = NULL,
  vr = NULL,
  Vmax = NULL,
  Vmin = NULL
)

Arguments

xi

state at the inflection point

vr

rate of change

Vmax

maximum terminal value

Vmin

minimum terminal value

Returns

list(xi,vr,Vmax,Vmin)

Method set_V_logistic_args()

Set V logistic arguments

Usage

FiniteDifference$set_V_logistic_args(
  xi = NULL,
  vr = NULL,
  Vmax = NULL,
  Vmin = NULL
)

Arguments

xi

state at the inflection point

vr

rate of change

Vmax

maximum terminal value

Vmin

minimum terminal value

Returns

list(xi,vr,Vmax,Vmin)

Method set_V_transcendental_args()

Set V transcendental arguments

Usage

FiniteDifference$set_V_transcendental_args(
  xo = NULL,
  xi = NULL,
  xm = NULL,
  Vmax = NULL,
  Vmin = NULL
)

Arguments

xo

state at the intercept

xi

state at the inflection point

xm

state at the maximum

Vmax

maximum terminal value

Vmin

minimum terminal value

Returns

list(xo,xi,xm,Vmax,Vmin)

Method set_V_yieldindex_args()

Set V yield index arguments

Usage

FiniteDifference$set_V_yieldindex_args(
  xo = NULL,
  xi = NULL,
  xm = NULL,
  Vmax = NULL,
  Vmin = NULL
)

Arguments

xo

state at the intercept

xi

state at the inflection point

xm

state at the maximum

Vmax

maximum terminal value

Vmin

minimum terminal value

Returns

list(xo,xi,xm,Vmax,Vmin)

Method set_V_args()

Set V arguments by position in list

Usage

FiniteDifference$set_V_args(
  Ix = NULL,
  name = NULL,
  v1 = NULL,
  v2 = NULL,
  v3 = NULL,
  v4 = NULL,
  v5 = NULL
)

Arguments

Ix

index of terminal value

name

name of terminal value

v1

first parameter

v2

second parameter

v3

third parameter

v4

fourth parameter

v5

fifth parameter

Returns

list(Ix,name,v1,v2,v3,v4,v5)

Method set_V_info()

Set information for terminal value

Usage

FiniteDifference$set_V_info(Ix = NULL, name = NULL)

Arguments

Ix

index of terminal value

name

name of terminal value

Returns

list(Ix,name,names,text)

Method set_plot_info()

Set information for plotting

Usage

FiniteDifference$set_plot_info(
  type = NULL,
  fontfamily = NULL,
  fontsize = NULL,
  fileformat = NULL,
  filewidth = NULL,
  fileheight = NULL,
  theme = NULL,
  opaque = NULL,
  walls = NULL,
  floor = NULL,
  labels = NULL,
  who = NULL
)

Arguments

type

= 1 and 2 for 2D, 3 and 4 for 3D

fontfamily

font family for plot labels

fontsize

font size for plot labels

fileformat

'png' or 'svg'

filewidth

pixel width of 2D plot, pixel width and height of 3D plot

fileheight

pixel height of 2D plot

theme

'light' or 'dark'

opaque

transparent to opaque background 0.0<=opaque<=1.0

walls

3D walls TRUE or FALSE

floor

3D floor TRUE or FALSE

labels

title and parameters TRUE or FALSE

who

identifier for object sending the parameters

Returns

list(type,font,file,theme,3D)

Method get_all()

Get all arguments and parameters

Usage

FiniteDifference$get_all()

Returns

list(oup_params,x_stoch_args,V_linear_args,V_stepped_args,V_kinked_args,V_butterfly_args,V_mitscherlich_args,V_gompertz_args,V_logistic_args,V_transcendental_args,V_yieldindex_args,plot_info)

Method get_oup_params()

Get OUP parameters

Usage

FiniteDifference$get_oup_params()

Returns

list(rho,mu,sigma)

Method get_x_stoch_args()

Get x as a stochastic state and its arguments

Usage

FiniteDifference$get_x_stoch_args()

Returns

list(s,x,V,r,phi,theta,skip,ds,dx)

Method get_V_linear_args()

Get V linear arguments

Usage

FiniteDifference$get_V_linear_args()

Returns

list(xo,vs)

Method get_V_degenerate_args()

Get V degenerate arguments

Usage

FiniteDifference$get_V_degenerate_args()

Returns

list(xo,Vmax,Vmin)

Method get_V_stepped_args()

Get V stepped arguments

Usage

FiniteDifference$get_V_stepped_args()

Returns

list(xo,vs,Vmax,Vmin)

Method get_V_kinked_args()

Get V kinked arguments

Usage

FiniteDifference$get_V_kinked_args()

Returns

list(xo,vs,Vmax,Vmin)

Method get_V_butterfly_args()

Get V butterfly arguments

Usage

FiniteDifference$get_V_butterfly_args()

Returns

list(xo,xm,vs,Vmax,Vmin)

Method get_V_mitscherlich_args()

Get V mitscherlich arguments

Usage

FiniteDifference$get_V_mitscherlich_args()

Returns

list(xo,vr,Vmax,Vmin)

Method get_V_gompertz_args()

Get V gompertz arguments

Usage

FiniteDifference$get_V_gompertz_args()

Returns

list(xi,vr,Vmax,Vmin)

Method get_V_logistic_args()

Get V logistic arguments

Usage

FiniteDifference$get_V_logistic_args()

Returns

list(xi,vr,Vmax,Vmin)

Method get_V_transcendental_args()

Get V transcendental arguments

Usage

FiniteDifference$get_V_transcendental_args()

Returns

list(xo,xi,xm,Vmax,Vmin)

Method get_V_yieldindex_args()

Get V yield index arguments

Usage

FiniteDifference$get_V_yieldindex_args()

Returns

list(xo,xi,xm,Vmax,Vmin)

Method get_V_args()

Get V arguments by index or name

Usage

FiniteDifference$get_V_args(Ix = NULL, name = NULL)

Arguments

Ix

index of terminal value

name

name of terminal value

Returns

list(xo,xi,xm,vs,vr,Vmax,Vmin)

Method get_V_info()

Get information for terminal values

Usage

FiniteDifference$get_V_info()

Returns

list(Ix,names)

Method get_plot_info()

Get information for plotting options

Usage

FiniteDifference$get_plot_info()

Returns

list(type,font,file,theme,3D,labels)

Method get_plot_colors()

Get colors for plotting

Usage

FiniteDifference$get_plot_colors()

Returns

list(red,ylw,grn,cyn,blu,mgn,gry,background,font,reverse)

Method axes_x_stoch()

Scale axes for x stochastic arguments

Usage

FiniteDifference$axes_x_stoch()

Returns

NULL

Method undo_clear()

Clear undo list and save current arguments to list

Usage

FiniteDifference$undo_clear()

Returns

1

Method undo_save()

Save current arguments to undo list

Usage

FiniteDifference$undo_save()

Returns

number of argument sets

Method undo_undo()

Replace current arguments from undo list

Usage

FiniteDifference$undo_undo()

Returns

c(index of this argument set, number of argument sets)

Method Drift()

Calculate, plot and return drifts

Usage

FiniteDifference$Drift(x = NULL, rho = NULL, mu = NULL, plotit = TRUE)

Arguments

x

vector of n states

rho

rate parameter 0<=<rho<inf

mu

location parameter -inf<mu<inf

plotit

TRUE or FALSE

Returns

list(g(1xn))

Method Diffusion()

Calculate, plot and return diffusions

Usage

FiniteDifference$Diffusion(x = NULL, sigma = NULL, plotit = TRUE)

Arguments

x

vector of n states

sigma

scale parameter -inf<sigma<inf

plotit

TRUE or FALSE

Returns

list(h2(1xn))

Method TerminalValue_Linear()

Create and plot linear terminal values

Usage

FiniteDifference$TerminalValue_Linear(
  x = NULL,
  xo = NULL,
  vs = NULL,
  plotit = TRUE
)

Arguments

x

vector of n states

xo

state at the intercept

vs

slope

plotit

TRUE or FALSE

Returns

list(V(1xn))

Method TerminalValue_Degenerate()

Create and plot degenerate terminal values

Usage

FiniteDifference$TerminalValue_Degenerate(
  x = NULL,
  xo = NULL,
  Vmax = NULL,
  Vmin = NULL,
  plotit = TRUE
)

Arguments

x

vector of n states

xo

state at the spike

Vmax

maximum terminal value

Vmin

minimum terminal value

plotit

TRUE or FALSE

Returns

list(V(1xn))

Method TerminalValue_Stepped()

Create and plot stepped terminal values

Usage

FiniteDifference$TerminalValue_Stepped(
  x = NULL,
  xo = NULL,
  vs = NULL,
  Vmax = NULL,
  Vmin = NULL,
  plotit = TRUE
)

Arguments

x

vector of n states

xo

state at the step

vs

direction of the step

Vmax

maximum terminal value

Vmin

minimum terminal value

plotit

TRUE or FALSE

Returns

list(V(1xn))

Method TerminalValue_Kinked()

Create and plot kinked terminal values

Usage

FiniteDifference$TerminalValue_Kinked(
  x = NULL,
  xo = NULL,
  vs = NULL,
  Vmax = NULL,
  Vmin = NULL,
  plotit = TRUE
)

Arguments

x

vector of n states

xo

state at the kink

vs

slope

Vmax

maximum terminal value

Vmin

minimum terminal value

plotit

TRUE or

Returns

list(V(1xn))

Method TerminalValue_Butterfly()

Create and plot butterfly terminal values

Usage

FiniteDifference$TerminalValue_Butterfly(
  x = NULL,
  xo = NULL,
  xm = NULL,
  vs = NULL,
  Vmax = NULL,
  Vmin = NULL,
  plotit = TRUE
)

Arguments

x

vector of n states

xo

state at the left wing

xm

state at the right wing

vs

slope

Vmax

maximum terminal value

Vmin

minimum terminal value

plotit

TRUE or

Returns

list(V(1xn))

Method TerminalValue_Mitscherlich()

Create and plot Mitscherlich terminal values

Usage

FiniteDifference$TerminalValue_Mitscherlich(
  x = NULL,
  xo = NULL,
  vr = NULL,
  Vmax = NULL,
  Vmin = NULL,
  plotit = TRUE
)

Arguments

x

vector of n states

xo

state at the intercept

vr

rate of change

Vmax

maximum terminal value

Vmin

minimum terminal value

plotit

TRUE or FALSE

Returns

list(V(1xn))

Method TerminalValue_Gompertz()

Create and plot Gompertz terminal values

Usage

FiniteDifference$TerminalValue_Gompertz(
  x = NULL,
  xi = NULL,
  vr = NULL,
  Vmax = NULL,
  Vmin = NULL,
  plotit = TRUE
)

Arguments

x

vector of n states

xi

state at the inflection point

vr

rate of change

Vmax

maximum terminal value

Vmin

minimum terminal value

plotit

TRUE or FALSE

Returns

list(V(1xn))

Method TerminalValue_Logistic()

Create and plot Logistic terminal values

Usage

FiniteDifference$TerminalValue_Logistic(
  x = NULL,
  xi = NULL,
  vr = NULL,
  Vmax = NULL,
  Vmin = NULL,
  plotit = TRUE
)

Arguments

x

vector of n states

xi

state at the inflection point

vr

rate of change

Vmax

maximum terminal value

Vmin

minimum terminal value

plotit

TRUE or FALSE

Returns

list(V(1xn))

Method TerminalValue_Transcendental()

Create and plot Transcendental terminal values

Usage

FiniteDifference$TerminalValue_Transcendental(
  x = NULL,
  xo = NULL,
  xi = NULL,
  xm = NULL,
  Vmax = NULL,
  Vmin = NULL,
  plotit = TRUE
)

Arguments

x

vector of n states

xo

state at the intercept

xi

state at the inflection point

xm

state at the maximum

Vmax

maximum terminal value

Vmin

minimum terminal value

plotit

TRUE or FALSE

Returns

list(V(1xn))

Method TerminalValue_YieldIndex()

Create and plot Yield Index terminal values

Usage

FiniteDifference$TerminalValue_YieldIndex(
  x = NULL,
  xo = NULL,
  xi = NULL,
  xm = NULL,
  Vmax = NULL,
  Vmin = NULL,
  plotit = TRUE
)

Arguments

x

vector of n states

xo

state at the intercept

xi

state at the inflection point

xm

state at the maximum

Vmax

maximum terminal value

Vmin

minimum terminal value

plotit

TRUE or FALSE

Returns

list(V(1xn))

Method TerminalValue()

Retrieves and plots terminal values

Usage

FiniteDifference$TerminalValue(Ix = NULL, name = NULL, plotit = TRUE)

Arguments

Ix

index number or name of terminal values

name

name of terminal values

plotit

TRUE or FALSE

Returns

list(V(1xn))

Method Option()

Calculate and plot option prices

Usage

FiniteDifference$Option(
  s = NULL,
  x = NULL,
  V = NULL,
  rho = NULL,
  mu = NULL,
  sigma = NULL,
  r = NULL,
  theta = NULL,
  skip = NULL,
  plotit = TRUE
)

Arguments

s

vector of m times

x

vector of n states

V

vector of n terminal values

rho

rate parameter 0<=rho<inf

mu

location parameter -inf<mu<inf

sigma

scale parameter -inf<sigma<inf

r

discount rate 0<=r<inf

theta

weight of current time in time stepping 0.5<=theta<=1

skip

divide the time interval and report every skip result

plotit

TRUE or FALSE

Returns

list(O(mxn))

Method OptionEnvelope()

Calculate and plot the envelope of option prices

Usage

FiniteDifference$OptionEnvelope(
  x = NULL,
  V = NULL,
  rho = NULL,
  mu = NULL,
  sigma = NULL,
  r = NULL,
  theta = NULL,
  skip = NULL,
  plotit = TRUE
)

Arguments

x

vector of n states

V

vector of n terminal values

rho

rate parameter 0<=rho<inf

mu

location parameter -inf<mu<inf

sigma

scale parameter -inf<sigma<inf

r

discount rate 0<=r<inf

theta

weight of current time in time stepping 0.5<=theta<=1

skip

divide the time interval and report every skip result

plotit

TRUE or FALSE

Returns

list(Ohat(1xn))

Method DecisionThreshold()

Calculate and plot the decision threshold

Usage

FiniteDifference$DecisionThreshold(
  x = NULL,
  V = NULL,
  rho = NULL,
  mu = NULL,
  sigma = NULL,
  r = NULL,
  phi = NULL,
  theta = NULL,
  skip = NULL,
  plotit = TRUE
)

Arguments

x

vector of n states

V

vector of n terminal values

rho

rate parameter 0<=rho<inf

mu

location parameter -inf<mu<inf

sigma

scale parameter -inf<sigma<inf

r

discount rate 0<=r<inf

phi

search direction for exit or entry options

theta

weight of current time in time stepping 0.5<=theta<=1

skip

divide the time interval and report every skip result

plotit

TRUE or FALSE

Returns

list(k,Ohat)

Method PlotDrift()

Plot drifts

Usage

FiniteDifference$PlotDrift(
  title = NULL,
  xaxis = NULL,
  yaxis = NULL,
  xbeg = NULL,
  xend = NULL
)

Arguments

title

text for plot title

xaxis

text for x-axis label

yaxis

text for y-axis label

xbeg

begin value for state axis

xend

end value for state axis

Returns

plot

Method PlotDiffusion()

Plot diffusions

Usage

FiniteDifference$PlotDiffusion(
  type = NULL,
  title = NULL,
  xaxis = NULL,
  yaxis = NULL,
  xbeg = NULL,
  xend = NULL
)

Arguments

type

= 2 and 3 for 2D

title

text for plot title

xaxis

text for x-axis label

yaxis

text for y-axis label

xbeg

begin value for state axis

xend

end value for state axis

Returns

plot

Method PlotTerminalValue()

Plot terminal values

Usage

FiniteDifference$PlotTerminalValue(
  title = NULL,
  xaxis = NULL,
  yaxis = NULL,
  xbeg = NULL,
  xend = NULL
)

Arguments

title

text for plot title

xaxis

text for x-axis label

yaxis

text for y-axis label

xbeg

begin value for state axis

xend

end value for state axis

Returns

plot

Method PlotOption()

Plot options

Usage

FiniteDifference$PlotOption(
  type = NULL,
  title = NULL,
  xaxis = NULL,
  yaxis = NULL,
  zaxis = NULL,
  sbeg = NULL,
  send = NULL,
  xbeg = NULL,
  xend = NULL
)

Arguments

type

= 2 for 2D, 3, 4 and 5 for 3D

title

text for plot title

xaxis

text for x-axis label

yaxis

text for y-axis label

zaxis

text for z-axis label

sbeg

begin value for time axis

send

end value for time axis

xbeg

begin value for state axis

xend

end value for state axis

Returns

plot

Method PlotOptionEnvelope()

Plot the option envelope

Usage

FiniteDifference$PlotOptionEnvelope(
  type = NULL,
  title = NULL,
  xaxis = NULL,
  yaxis = NULL,
  zaxis = NULL,
  sbeg = NULL,
  send = NULL,
  xbeg = NULL,
  xend = NULL
)

Arguments

type

= 3 for 2D, 4 for 3D

title

text for plot title

xaxis

text for x-axis label

yaxis

text for y-axis label

zaxis

text for z-axis label

sbeg

begin value for time axis

send

end value for time axis

xbeg

begin value for state axis

xend

end value for state axis

Returns

plot

Method PlotDecisionThreshold()

Plot the decision threshold

Usage

FiniteDifference$PlotDecisionThreshold(
  title = NULL,
  xaxis = NULL,
  yaxis = NULL,
  xbeg = NULL,
  xend = NULL
)

Arguments

title

text for plot title

xaxis

text for x-axis label

yaxis

text for y-axis label

xbeg

begin value for state axis

xend

end value for state axis

Returns

plot

Examples

## ------------------------------------------------
## Method `FiniteDifference$new`
## ------------------------------------------------

  FD <- FiniteDifference$new()
  FD$TerminalValue_Mitscherlich(plotit=FALSE)
  FD$DecisionThreshold()

[Package GregsOUPR6 version 1.3.5.0 Index]