MaximumLikelihood {GregsOUPR6}	R Documentation

R6 Class implementing Maximum Likelihood estimation

Description

Maximum Likelihood estimation of the Ornstein-Uhlenbeck Process, using a robust estimation algorithm and with testing for simple hypotheses.

Methods:

rho, mu and sigma random
  LogLikelihood
  Estimates
  GoodnessOfFit
  LikelihoodRatioTest

Plots:

  PlotTimeSeries
  PlotEstimates

Arguments of functions:

  All arguments are optional in all functions.
  rho:    rate parameter 0<=rho<inf
  mu:     location parameter -inf<mu<inf
  sigma:  scale parameter -inf<sigma<inf
  tau:    vector of observed times -inf<tau<inf
  z:      vector of observed states -inf<z<inf
  df:     data frame containing columns for tau and z
  taucol: index of a column containing tau
  zcol:   index of a column containing z
  rhor:   constant to fix the rate parameter 0<=rhor<inf
  mur:    constant to fix the location parameter -inf<mur<inf
  sigmar: constant to fix scale parameter -inf<sigmar<inf
  rhos:   starting value for the rate parameter 0<=rhos<inf
  mus:    starting value for the location parameter -inf<mus<inf
  sigmas: starting value the scale parameter -inf<sigmas<inf
  lnLu:   unrestricted log likelihood -inf<lnLu<=0
  alphau: identifies distribution of lnLu 1<=alphau<=2
  lnLr:   restricted log likelihood -inf<lnLr<=lnLu
  alpha: identifies distribution of lnLu 1<=alpha<=2
  m1:     number of observations 0<m1=m-1

Using the methods:

Demonstration scripts are in files in the 'demo' directory. Identify a formula and in the console type:

  demo(ML_FormulaName), or
  demo(ML_PlotFormulaName).

Discussion:

The Nelder-Mead algorithm is used to maximize a Log Likelihood derived from the Transition Density of the Ornstein-Uhlenbeck Process. The original algorithm is designed to minimize and has been called the amoeba algorithm. If spread across a bumpy surface, the amoeba pulls itself over bumps and out of hollows and flows down to the lowest level. There it contracts around its center. If spread across a flat surface, the amoeba shrinks without flowing. Once the amoeba contracts or shrinks, it is spread across the surface opposite from where is came. Again it flows, contracts and shrinks. Spreading, flowing, contracting and shrinking continue until the amoeba contracts around the same point or shrinks at the same level at least twice.

In this implementation, the Nelder-Mead algorithm has been modified in five ways.

It is modified to maximize;

It can set parameters rho, mu and/or sigma to constants for simple hypothesis tests;

It has a tie-breaking condition to prevent cycling or freezing;

It can accelerate over long distances for searching flat likelihoods;

It checks for log likelihoods greater than zero to prevent errors in case the data is faked or artificially smoothed.

Once the unrestricted estimates are found, the arguments for restricting the parameters can be set for simple hypothesis tests. By default, starting values are calculated from the data and are then updated automatically. The arguments for starting values should seldom be needed.

Data are entered as a data frame with columns for tau, a vector of times, and z, a vector of states. Arguments taucol and zcol are columns to extract from a data frame containing many columns.

To read a csv file into a data frame, create a MaximumLikelihood object, estimate parameters and test the goodness-of-fit:

  df <- read.csv("data/ML_OUP_SimulatedData.csv")
  ML <- MaximumLikelihood$new()
  ML$Estimates(df)
  ML$GoodnessOfFit()

In this example, the 'data' directory is one level below the working directory. The taucol and zcol arguments are optional and will default to 1 and 2. Continue on by restricting a parameter and performing a likelihood ratio test:

  ML$Estimates(rhor=0.5)
  ML$LikelihoodRatioTest()

The available data sets are documented. To see the documentation, type:

  ?ML_

and then select from the drop-down list.

Examples in R6 don't work the same way as other R modules. There is only one example for an R6 object, not one for each function in the object. To run examples, the devtools::run_examples() works, but the R command example("MaximumLikelihood") doesn't. You can copy commands to the clipboard, paste into the console and press Enter. Examples in this help and a simple example at the bottom can be run in this way. A better alternative is demo().

A better alternative is demo(), but this works sometimes, sometimes not. If you don't see the demos for this package, go to Files and demo. You will see the demo names and then type something like demo(ML_Estimates).

Methods

Public methods

MaximumLikelihood$new()

MaximumLikelihood$set_oup_params()

MaximumLikelihood$set_oup_params_restr()

MaximumLikelihood$set_oup_params_start()

MaximumLikelihood$set_oup_stats()

MaximumLikelihood$set_timeseries()

MaximumLikelihood$set_timeseries_info()

MaximumLikelihood$set_plot_info()

MaximumLikelihood$get_all()

MaximumLikelihood$get_oup_params()

MaximumLikelihood$get_oup_params_unrestr()

MaximumLikelihood$get_oup_params_restr()

MaximumLikelihood$get_oup_params_start()

MaximumLikelihood$get_oup_stats()

MaximumLikelihood$get_timeseries()

MaximumLikelihood$get_timeseries_info()

MaximumLikelihood$get_plot_info()

MaximumLikelihood$get_plot_colors()

MaximumLikelihood$LogLikelihood()

MaximumLikelihood$Estimates()

MaximumLikelihood$GoodnessOfFit()

MaximumLikelihood$LikelihoodRatioTest()

MaximumLikelihood$PlotTimeSeries()

MaximumLikelihood$PlotEstimates()

Method new()

Create a MaximumLikelihood object

Usage

MaximumLikelihood$new(OUP = NULL)

Arguments

OUP

pointer set by the OUProcess object

Returns

A new MaximumLikelihood object

Examples

  ML <- MaximumLikelihood$new()
  ML$Estimates()
  ML$GoodnessOfFit()
  ML$Estimates(rhor=0.5,mur=-15,sigmar=15)
  ML$LikelihoodRatioTest()

Method set_oup_params()

Set OUP parameters

Usage

MaximumLikelihood$set_oup_params(
  rho = NULL,
  mu = NULL,
  sigma = NULL,
  who = NULL
)

Arguments

rho

rate parameter 0<=rho<inf

mu

location parameter -inf<mu<inf

sigma

scale parameter -inf<sigma<inf

who

identifier for object sending the parameters

Returns

list(rho,mu,sigma) Set OUP parameter restrictions

Method set_oup_params_restr()

Usage

MaximumLikelihood$set_oup_params_restr(rhor = NULL, mur = NULL, sigmar = NULL)

Arguments

rhor

rate parameter 0<=rhor<inf

mur

location parameter -inf<mur<inf

sigmar

scale parameter -inf<sigmar<inf

Returns

list(rhor,mur,sigmar) Set OUP parameter starting values

Method set_oup_params_start()

Usage

MaximumLikelihood$set_oup_params_start(rhos = NULL, mus = NULL, sigmas = NULL)

Arguments

rhos

rate parameter 0<=rhos<inf

mus

location parameter -inf<mus<inf

sigmas

scale parameter -inf<sigmas<inf

Returns

list(rhos,mus,sigmas)

Method set_oup_stats()

Set OUP statistics for estimation and hypothesis tests

Usage

MaximumLikelihood$set_oup_stats(
  lnLu = NULL,
  alphau = NULL,
  lnLr = NULL,
  alphar = NULL,
  m1 = NULL
)

Arguments

lnLu

unrestricted log likelihood -inf<lnLu<=0

alphau

identifies the distribution of lnLu, 0.5<=alphau<=1

lnLr

restricted log likelihood -inf<lnLr<=lnLu

alphar

identifies the distribution of lnLr, 0.5<=alphar<=1

m1

number of observations 0<m1

Returns

list(lnLu,alphau,lnLr,alphar,m1)

Method set_timeseries()

Set time series data for time tau and state z

Usage

MaximumLikelihood$set_timeseries(df = NULL, taucol = NULL, zcol = NULL)

Arguments

df

data frame containing columns for tau and z

taucol

index of a column containing tau

zcol

index a column containing z

Returns

dataframe(tau,z)

Method set_timeseries_info()

Set information for plotting times series and estimates

Usage

MaximumLikelihood$set_timeseries_info(
  tbeg = NULL,
  tend = NULL,
  dataname = NULL,
  timename = NULL,
  statename = NULL,
  estimation = NULL
)

Arguments

tbeg

begin value for time axis

tend

end value for time axis

dataname

name for the data

timename

name for time

statename

name for state

estimation

description of estimation

Returns

list(tbeg,Ixbeg,tend,Ixend,datename,timename,statename,estimation)

Method set_plot_info()

Set information for plotting

Usage

MaximumLikelihood$set_plot_info(
  fontfamily = NULL,
  fontsize = NULL,
  fileformat = NULL,
  filewidth = NULL,
  fileheight = NULL,
  theme = NULL,
  opaque = NULL,
  labels = NULL,
  who = NULL
)

Arguments

fontfamily

font family for plot labels

fontsize

font size for plot labels

fileformat

'png' or 'svg'

filewidth

pixel width of 2D plot, pixel width and height of 3D plot

fileheight

pixel height of 2D plot

theme

'light' or 'dark'

opaque

transparent to opaque background 0.0<=opaque<=1.0

labels

title and parameters TRUE or FALSE

who

identifier for object sending the parameters

Returns

list(font,file,theme)

Method get_all()

Get all arguments, time series and information

Usage

MaximumLikelihood$get_all()

Returns

list(oup_params,oup_params_unrestr,oup_params_restr,oup_params_start,oup_stats,timeseries,timeseries_info,plot_info)

Method get_oup_params()

Get OUP parameters

Usage

MaximumLikelihood$get_oup_params()

Returns

list(rho,mu,sigma)

Method get_oup_params_unrestr()

Get OUP parameters unrestricted estimates

Usage

MaximumLikelihood$get_oup_params_unrestr()

Returns

list(rhohat,muhat,sigmahat)

Method get_oup_params_restr()

Get OUP parameter restrictions

Usage

MaximumLikelihood$get_oup_params_restr()

Returns

list(rhor,mur,sigmar)

Method get_oup_params_start()

Get OUP parameter starting values

Usage

MaximumLikelihood$get_oup_params_start()

Returns

list(rhos,mus,sigmas)

Method get_oup_stats()

Get OUP statistics for estimation and hypothesis tests

Usage

MaximumLikelihood$get_oup_stats()

Returns

list(lnLu,alphau,lnLr,alphar,m1)

Method get_timeseries()

Get time series data for time tau and state z

Usage

MaximumLikelihood$get_timeseries()

Returns

dataframe(tau,z)

Method get_timeseries_info()

Get information for times series

Usage

MaximumLikelihood$get_timeseries_info()

Returns

list(tbeg,Ixbeg,tend,Ixend,datename,timename,statename,estimation)

Method get_plot_info()

Get information for plotting

Usage

MaximumLikelihood$get_plot_info()

Returns

list(font,file,theme,labels)

Method get_plot_colors()

Get colors for plotting

Usage

MaximumLikelihood$get_plot_colors()

Returns

list(red,ylw,grn,cyn,blu,mgn,gry,background,font,reverse)

Method LogLikelihood()

Calculate the log likelihood of estimates

Usage

MaximumLikelihood$LogLikelihood(
  rho = NULL,
  mu = NULL,
  sigma = NULL,
  df = NULL,
  taucol = NULL,
  zcol = NULL
)

Arguments

rho

rate parameter 0<=rho<inf

mu

location parameter -inf<mu<inf

sigma

scale parameter -inf<sigma<inf

df

data frame containing columns for tau and z

taucol

index of a column containing tau

zcol

index of a column containing z

Returns

list(lnL)

Method Estimates()

Calculate unrestricted or restricted maximum likelihood estimates

Usage

MaximumLikelihood$Estimates(
  df = NULL,
  taucol = NULL,
  zcol = NULL,
  rhor = NULL,
  mur = NULL,
  sigmar = NULL,
  rhos = NULL,
  mus = NULL,
  sigmas = NULL,
  plotit = TRUE
)

Arguments

df

data frame containing columns for tau and z

taucol

index of a column containing tau

zcol

index of a column containing z

rhor

constant to fix the rate parameter 0<=rhor<inf

mur

constant to fix the location parameter -inf<mur<inf

sigmar

constant to fix scale parameter -inf<sigmar<inf

rhos

starting value for the rate parameter 0<=rhos<inf

mus

starting value for the location parameter -inf<mus<inf

sigmas

starting value the scale parameter -inf<sigmas<inf

plotit

TRUE or FALSE

Returns

list(rhohat,muhat,sigmahat,lnLu,ku,alphau,m1) or list(rhor,mur,sigmar,lnLr,kr,alphar,m1)

Method GoodnessOfFit()

Calculate the goodness of fit compared with invariant and scaled brownian motion estimates

Usage

MaximumLikelihood$GoodnessOfFit(
  rho = NULL,
  mu = NULL,
  sigma = NULL,
  df = NULL,
  taucol = NULL,
  zcol = NULL
)

Arguments

rho

rate parameter 0<=rho<inf

mu

location parameter -inf<mu<inf

sigma

scale parameter -inf<sigma<inf

df

data frame containing columns for tau and z

taucol

index of a column containing tau

zcol

index a column containing z

Returns

list(Inv,SBM) with Inv=list(R2,Pval) and SBM=list(R2,Pval)

Method LikelihoodRatioTest()

Test for significant difference between two likelihoods

Usage

MaximumLikelihood$LikelihoodRatioTest(
  lnLu = NULL,
  lnLr = NULL,
  alphau = NULL,
  m1 = NULL
)

Arguments

lnLu

unrestricted log likelihood -inf<lnLu<=0

lnLr

restricted log likelihood -inf<lnLr<=lnLu

alphau

identifies distribution of LnLu 0.5<=alphau<=1

m1

number of observations 0<m1=m-1

Returns

list(R2,Pval)

Method PlotTimeSeries()

Plot time series

Usage

MaximumLikelihood$PlotTimeSeries(
  df = NULL,
  taucol = NULL,
  zcol = NULL,
  tbeg = NULL,
  tend = NULL,
  title = NULL,
  xaxis = NULL,
  yaxis = NULL
)

Arguments

df

data frame containing columns for tau and z

taucol

index of a column containing tau

zcol

index a column containing z

tbeg

begin value for time axis

tend

end value for time axis

title

text for plot title

xaxis

text for x-axis label

yaxis

text for y-axis label

Returns

plot

Method PlotEstimates()

Plot time series with estimated means and variances

Usage

MaximumLikelihood$PlotEstimates(
  tbeg = NULL,
  tend = NULL,
  title = NULL,
  xaxis = NULL,
  yaxis = NULL
)

Arguments

tbeg

begin value for time axis

tend

end value for time axis

title

text for plot title

xaxis

text for x-axis label

yaxis

text for y-axis label

Returns

plot

Examples

## ------------------------------------------------
## Method `MaximumLikelihood$new`
## ------------------------------------------------

  ML <- MaximumLikelihood$new()
  ML$Estimates()
  ML$GoodnessOfFit()
  ML$Estimates(rhor=0.5,mur=-15,sigmar=15)
  ML$LikelihoodRatioTest()

[Package GregsOUPR6 version 1.3.5.0 Index]