OUP_ObservationInterval {GregsOUPR6}	R Documentation

Observation intervals and the Ornstein-Uhlenbeck Process

Description

Monte-Carlo simulation to demonstrate the effect of the observation interval on the rate of convergence, rho, and the scale, sigma.

Format

csv file with 201 rows and 4 columns

Details

year 1: time variable in annual increments

z rho 0.5: sample paths with annual time increments and a rate of convergence of 0.5

year 0.05: time variable in increments of 1/20 year

z rho 10: sample paths with time increments of 1/20 year and a rate of convergence of 10

This data has two sample paths. Within a sample path, observation intervals are of equal length. Between sample paths the observation intervals are of unequal lengths. Otherwise, the sample paths have the same shocks and give equivalent estimates of the parameters.

Both sets of estimates have the same rho(t-s), where t-s is the observation interval. A common example is annual versus daily intervals; t-s=1 and rho=1/2 is equivalent to t-s=1/365 and rho=365/2. Location parameter, mu, is the same in both sets of estimates, but scale parameter, sigma, is larger if rho is larger because the asymptotic variance, sigma^2/2rho, is the same in both sets of estimates.

[Package GregsOUPR6 version 1.3.5.0 Index]