OUP_SampleSize {GregsOUPR6}	R Documentation

Sample sizes for the Ornstein-Uhlenbeck Process

Description

Monte-Carlo simulation to demonstrate the effect of sample size on hypothesis tests for the Ornstein-Uhlenbeck Process.

Format

csv file with 4736 rows and 6 columns

Details

year: time variable in annual increments for all sample paths

xsmall: sample with 37 measurements or 36 observations

small: sample with 2 x 37 measurements or 73 observations

medium: sample with 8 x 37 measurements or 295 observations

large: sample with 32 x 37 measurements or 1183 observations

xlarge: sample with 128 x 37 measurements or 4735 observations

This data starts with a very small sample and then replicates it. Larger samples contain no more information than smaller samples and hypothesis tests for the Ornstein-Uhlenbeck Process do not change with the sample size. The usual Chi^2 tests, however, will reject all null hypotheses for medium to large sample sizes.

If larger samples contain more information, hypothesis tests for the Ornstein-Uhlenbeck Process will reject more null hypotheses, but not all hypotheses as do Chi^2 tests.

[Package GregsOUPR6 version 1.3.5.0 Index]